v0.0.2 - Add reauth to world list

+++ game-server/src/main/java/com/aaemu/game/service/enums/unit/UnitGender.java
@@ -10,10 +10,10 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 @Getter
 public enum UnitGender {
-    MALE(1),
-    FEMALE(2);
+    MALE((byte) 1),
+    FEMALE((byte) 2);
 
-    private final int id;
+    private final byte id;
 
     public static UnitGender get(byte id) {
         for (UnitGender unitGender : values()) {
